This moves some of the new enum groups introduced in #108 into distinct \<group> tags and removes the value= attributes in those groups, because the scripts don't allow redefining enums (even benignly) inside \<enums> tags.
I'm not certain what to do with the additions to ShaderType and they
have just been reverted, for the moment. Probably best is to use a
different group name (ShaderStages?) in a \<group> tag including all of
them.

Please make sure that after proposing a new PR addressing this, 'make'
executes and regenerates the headers, unchanged other than the revision
dates, without reporting any warnings or errors.